Squashed commit of the following: /save confirmation + readline
commit 0b6a3822a0e7ff321e734864787371a92719351c
Author: Frank Schwidom <schwidom@gmx.net>
Date:   Fri Jul 31 01:08:55 2026 +0200

    readline

commit 0ffff4d7ea80b53770fbe7ef868840d9fbf4077f
Author: Frank Schwidom <schwidom@gmx.net>
Date:   Fri Jul 31 00:42:46 2026 +0200

    /save confirmation

+++ chat.py
@@ -3,6 +3,11 @@ import os
 import sys
 from dataclasses import dataclass, field
 
+try:
+    import readline
+except ImportError:
+    readline = None
+
 from tool_base import Tool, run_with_tools
 
 
@@ -196,6 +201,12 @@ def _cmd_save(path: str, state: ChatState):
         "model": state.model,
         "messages": state.messages,
     }
+    if os.path.exists(path):
+        confirm = input(f"File '{path}' already exists. Overwrite? (y/n): ").lower()
+        if confirm != 'y':
+            print("Save aborted.")
+            return
+
     try:
         with open(path, "w", encoding="utf-8") as f:
             json.dump(data, f, indent=2, ensure_ascii=False)
